position states to redux. this commit includes many errors because tree data states have been changed but not fully transfered to redux
