+++ src/store.js
@@ -1,8 +1,10 @@
 import { configureStore } from '@reduxjs/toolkit'
-import userReducer from './components/userSlice.js'
+import userReducer from './slices/userSlice.js'
+import positionReducer from './slices/positionSlice.js'
 
 export default configureStore({
   reducer: {
-    user: userReducer
+    user: userReducer,
+    position: positionReducer
   }
 })
